Refactor to coreIssuance
