+++ test/contracts/core/modules/exchangeIssuanceModule.spec.ts
@@ -7,7 +7,7 @@ import * as setProtocolUtils from 'set-protocol-utils';
 import {
   Address,
   Bytes,
-  ExchangeInteractData,
+  ExchangeIssuanceParams,
   KyberTrade,
   ZeroExSignedFillOrder
 } from 'set-protocol-utils';
@@ -17,7 +17,7 @@ import ChaiSetup from '@utils/chaiSetup';
 import { BigNumberSetup } from '@utils/bigNumberSetup';
 import {
   CoreContract,
-  ExchangeIssueModuleContract,
+  ExchangeIssuanceModuleContract,
   SetTokenContract,
   SetTokenFactoryContract,
   StandardTokenMockContract,
@@ -28,7 +28,7 @@ import { ether } from '@utils/units';
 import { assertTokenBalanceAsync, expectRevertError } from '@utils/tokenAssertions';
 import { Blockchain } from '@utils/blockchain';
 import { DEFAULT_GAS, DEPLOYED_TOKEN_QUANTITY, KYBER_RESERVE_CONFIGURED_RATE } from '@utils/constants';
-import { LogExchangeIssue } from '@utils/contract_logs/exchangeIssueModule';
+import { LogExchangeIssue, LogExchangeRedeem } from '@utils/contract_logs/exchangeIssuanceModule';
 import { generateOrdersDataWithIncorrectExchange } from '@utils/orders';
 import { getWeb3 } from '@utils/web3Helper';
 
@@ -40,7 +40,7 @@ BigNumberSetup.configure();
 ChaiSetup.configure();
 const web3 = getWeb3();
 const Core = artifacts.require('Core');
-const ExchangeIssueModule = artifacts.require('ExchangeIssueModule');
+const ExchangeIssuanceModule = artifacts.require('ExchangeIssuanceModule');
 const { SetProtocolTestUtils: SetTestUtils, SetProtocolUtils: SetUtils } = setProtocolUtils;
 const blockchain = new Blockchain(web3);
 const setTestUtils = new SetTestUtils(web3);
@@ -48,18 +48,18 @@ const setUtils = new SetUtils(web3);
 const { expect } = chai;
 const { NULL_ADDRESS, ZERO } = SetUtils.CONSTANTS;
 
-contract('ExchangeIssueModule', accounts => {
+contract('ExchangeIssuanceModule', accounts => {
   const [
     contractDeployer,
     notExchangeIssueCaller,
     zeroExOrderMaker,
-    exchangeIssueCaller,
+    exchangeIssuanceCaller,
   ] = accounts;
 
   let core: CoreContract;
   let transferProxy: TransferProxyContract;
   let vault: VaultContract;
-  let exchangeIssueModule: ExchangeIssueModuleContract;
+  let exchangeIssuanceModule: ExchangeIssuanceModuleContract;
   let setTokenFactory: SetTokenFactoryContract;
 
   const coreWrapper = new CoreWrapper(contractDeployer, contractDeployer);
@@ -68,12 +68,12 @@ contract('ExchangeIssueModule', accounts => {
 
   before(async () => {
     ABIDecoder.addABI(Core.abi);
-    ABIDecoder.addABI(ExchangeIssueModule.abi);
+    ABIDecoder.addABI(ExchangeIssuanceModule.abi);
   });
 
   after(async () => {
     ABIDecoder.removeABI(Core.abi);
-    ABIDecoder.removeABI(ExchangeIssueModule.abi);
+    ABIDecoder.removeABI(ExchangeIssuanceModule.abi);
   });
 
   beforeEach(async () => {
@@ -82,11 +82,11 @@ contract('ExchangeIssueModule', accounts => {
     vault = await coreWrapper.deployVaultAsync();
     transferProxy = await coreWrapper.deployTransferProxyAsync();
     core = await coreWrapper.deployCoreAsync(transferProxy, vault);
-    exchangeIssueModule = await coreWrapper.deployExchangeIssueModuleAsync(
+    exchangeIssuanceModule = await coreWrapper.deployExchangeIssuanceModuleAsync(
       core,
       vault
     );
-    await coreWrapper.addModuleAsync(core, exchangeIssueModule.address);
+    await coreWrapper.addModuleAsync(core, exchangeIssuanceModule.address);
     setTokenFactory = await coreWrapper.deploySetTokenFactoryAsync(core.address);
 
     await coreWrapper.setDefaultStateAndAuthorizationsAsync(core, vault, transferProxy, setTokenFactory);
@@ -98,7 +98,7 @@ contract('ExchangeIssueModule', accounts => {
 
   describe('#exchangeIssue', async () => {
     let subjectCaller: Address;
-    let subjectExchangeInteractData: ExchangeInteractData;
+    let subjectExchangeIssuanceParams: ExchangeIssuanceParams;
     let subjectExchangeOrdersData: Bytes;
 
     let naturalUnit: BigNumber;
@@ -157,7 +157,7 @@ contract('ExchangeIssueModule', accounts => {
       await erc20Wrapper.approveTransfersAsync(
         [sentToken],
         transferProxy.address,
-        exchangeIssueCaller
+        exchangeIssuanceCaller
       );
       await erc20Wrapper.approveTransfersAsync(
         [secondComponent],
@@ -191,7 +191,7 @@ contract('ExchangeIssueModule', accounts => {
         );
 
       // Property:                Value                         | Property
-      subjectExchangeInteractData = {
+      subjectExchangeIssuanceParams = {
         setAddress:             exchangeIssueSetAddress,            // setAddress
         sentTokenExchangeIds:   exchangeIssueSentTokenExchangeIds,  // sentTokenExchangeIds
         sentTokens:             exchangeIssueSentTokens,            // sentToken
@@ -199,7 +199,7 @@ contract('ExchangeIssueModule', accounts => {
         quantity:               exchangeIssueQuantity,              // quantity
         receiveTokens:          exchangeIssueReceiveTokens,         // requiredComponents
         receiveTokenAmounts:    exchangeIssueReceiveTokenAmounts,   // requiredComponentAmounts
-      } as ExchangeInteractData;
+      } as ExchangeIssuanceParams;
 
       // Create Kyber trade for the second component, using ether(25) sentToken. Conversion rate pre set on snapshot
       const maxDestinationQuantity = exchangeIssueReceiveTokenAmounts[0];
@@ -241,7 +241,7 @@ contract('ExchangeIssueModule', accounts => {
 
       subjectExchangeOrdersData =
         subjectExchangeOrdersData || setUtils.generateSerializedOrders([zeroExOrder, kyberTrade]);
-      subjectCaller = exchangeIssueCaller;
+      subjectCaller = exchangeIssuanceCaller;
     });
 
     afterEach(async () => {
@@ -251,31 +251,31 @@ contract('ExchangeIssueModule', accounts => {
     });
 
     async function subject(): Promise<string> {
-      return exchangeIssueModule.exchangeIssue.sendTransactionAsync(
-        subjectExchangeInteractData,
+      return exchangeIssuanceModule.exchangeIssue.sendTransactionAsync(
+        subjectExchangeIssuanceParams,
         subjectExchangeOrdersData,
         { from: subjectCaller, gas: DEFAULT_GAS },
       );
     }
 
     it('mints the correct quantity of the set for the sender', async () => {
-      const existingBalance = await setToken.balanceOf.callAsync(exchangeIssueCaller);
+      const existingBalance = await setToken.balanceOf.callAsync(exchangeIssuanceCaller);
 
       await subject();
 
-      await assertTokenBalanceAsync(setToken, existingBalance.add(exchangeIssueQuantity), exchangeIssueCaller);
+      await assertTokenBalanceAsync(setToken, existingBalance.add(exchangeIssueQuantity), exchangeIssuanceCaller);
     });
 
     it('transfers the maker token amount from the maker, and returns change from Kyber', async () => {
-      const existingBalance = await sentToken.balanceOf.callAsync(exchangeIssueCaller);
-      await assertTokenBalanceAsync(sentToken, DEPLOYED_TOKEN_QUANTITY, exchangeIssueCaller);
+      const existingBalance = await sentToken.balanceOf.callAsync(exchangeIssuanceCaller);
+      await assertTokenBalanceAsync(sentToken, DEPLOYED_TOKEN_QUANTITY, exchangeIssuanceCaller);
 
       await subject();
 
       // TODO: Change from unused kyber source token is not being calculated correctly, off by 3 * 10 ** -26
       const expectedNewBalance = existingBalance.sub(totalSentToken)
                                                 .add(kyberTradeMakerTokenChange);
-      const newBalance = await sentToken.balanceOf.callAsync(exchangeIssueCaller);
+      const newBalance = await sentToken.balanceOf.callAsync(exchangeIssuanceCaller);
 
       await expect(newBalance.toPrecision(26)).to.be.bignumber.equal(expectedNewBalance.toPrecision(26));
     });
@@ -300,7 +300,7 @@ contract('ExchangeIssueModule', accounts => {
         exchangeIssueQuantity,
         exchangeIssueSentTokens,
         exchangeIssueSentTokenAmounts,
-        exchangeIssueModule.address
+        exchangeIssuanceModule.address
       );
 
       await SetTestUtils.assertLogEquivalence(formattedLogs, expectedLogs);
@@ -510,6 +510,284 @@ contract('ExchangeIssueModule', accounts => {
 
       after(async () => {
         exchangeIssueReceiveTokenAmounts = undefined;
+      });
+
+      it('should revert', async () => {
+        await expectRevertError(subject());
+      });
+    });
+  });
+
+  describe('#exchangeRedeem', async () => {
+    let subjectCaller: Address;
+    let subjectExchangeIssuanceParams: ExchangeIssuanceParams;
+    let subjectExchangeOrdersData: Bytes;
+
+    let naturalUnit: BigNumber;
+    let setToken: SetTokenContract;
+    let receiveToken: StandardTokenMockContract;
+
+    let totalReceiveToken: BigNumber;
+
+    let exchangeRedeemSetAddress: Address;
+    let exchangeRedeemQuantity: BigNumber;
+    let exchangeRedeemSentTokenExchanges: BigNumber[];
+    let exchangeRedeemSentTokens: Address[];
+    let exchangeRedeemSentTokenAmounts: BigNumber[];
+    let exchangeRedeemReceiveTokens: Address[];
+    let exchangeRedeemReceiveTokenAmounts: BigNumber[];
+
+    let zeroExOrder: ZeroExSignedFillOrder;
+    let zeroExOrderMakerTokenAmount: BigNumber;
+    let zeroExOrderTakerTokenAmount: BigNumber;
+    let kyberTrade: KyberTrade;
+    let kyberConversionRatePower: BigNumber;
+
+    beforeEach(async () => {
+      subjectCaller = exchangeIssuanceCaller;
+
+      await exchangeWrapper.deployAndAuthorizeZeroExExchangeWrapper(
+        core,
+        SetTestUtils.ZERO_EX_EXCHANGE_ADDRESS,
+        SetTestUtils.ZERO_EX_ERC20_PROXY_ADDRESS,
+        SetTestUtils.ZERO_EX_TOKEN_ADDRESS,
+        transferProxy
+      );
+      await exchangeWrapper.deployAndAuthorizeKyberNetworkWrapper(
+        core,
+        SetTestUtils.KYBER_NETWORK_PROXY_ADDRESS,
+        transferProxy
+      );
+
+      const firstComponent = erc20Wrapper.kyberReserveToken(SetTestUtils.KYBER_RESERVE_SOURCE_TOKEN_ADDRESS);
+      const secondComponent = await erc20Wrapper.deployTokenAsync(contractDeployer);
+      receiveToken = erc20Wrapper.kyberReserveToken(SetTestUtils.KYBER_RESERVE_DESTINATION_TOKEN_ADDRESS);
+
+      const componentTokens = [firstComponent, secondComponent];
+      const setComponentUnit = ether(4);
+      const componentAddresses = componentTokens.map(token => token.address);
+      const componentUnits = componentTokens.map(token => setComponentUnit);
+      naturalUnit = ether(2);
+
+      setToken = await coreWrapper.createSetTokenAsync(
+        core,
+        setTokenFactory.address,
+        componentAddresses,
+        componentUnits,
+        naturalUnit,
+      );
+
+      zeroExOrderMakerTokenAmount = zeroExOrderMakerTokenAmount || ether(4);
+      const kyberDestinationTokenQuantity = ether(2.56);
+      totalReceiveToken = zeroExOrderMakerTokenAmount.add(kyberDestinationTokenQuantity);
+
+      exchangeRedeemSetAddress = exchangeRedeemSetAddress || setToken.address;
+      exchangeRedeemQuantity = exchangeRedeemQuantity || ether(4);
+
+      exchangeRedeemSentTokenExchanges =
+        exchangeRedeemSentTokenExchanges || [SetUtils.EXCHANGES.KYBER, SetUtils.EXCHANGES.ZERO_EX];
+
+      exchangeRedeemSentTokens = exchangeRedeemSentTokens || [firstComponent.address, secondComponent.address];
+      exchangeRedeemSentTokenAmounts =
+        exchangeRedeemSentTokenAmounts || _.map(componentUnits, unit => unit
+          .mul(exchangeRedeemQuantity)
+          .div(naturalUnit)
+        );
+
+      exchangeRedeemReceiveTokens =
+        exchangeRedeemReceiveTokens || [receiveToken.address];
+
+      exchangeRedeemReceiveTokenAmounts =
+        exchangeRedeemReceiveTokenAmounts || [totalReceiveToken];
+
+      // Property:                Value                         | Property
+      subjectExchangeIssuanceParams = {
+        setAddress:             exchangeRedeemSetAddress,          // setAddress
+        sentTokenExchangeIds:     exchangeRedeemSentTokenExchanges,  // sentTokenExchangeIds
+        sentTokens:             exchangeRedeemSentTokens,          // sentToken
+        sentTokenAmounts:       exchangeRedeemSentTokenAmounts,    // sentTokenAmount
+        quantity:               exchangeRedeemQuantity,            // quantity
+        receiveTokens:          exchangeRedeemReceiveTokens,       // receiveTokens
+        receiveTokenAmounts:    exchangeRedeemReceiveTokenAmounts, // receiveTokenAmounts
+      } as ExchangeIssuanceParams;
+
+      // Create Kyber trade. Conversion rate pre set on snapshot
+      const sourceTokenQuantity = exchangeRedeemSentTokenAmounts[0];
+      const maxDestinationQuantity = kyberDestinationTokenQuantity;
+      const destinationTokenDecimals = (await receiveToken.decimals.callAsync()).toNumber();
+      const sourceTokenDecimals = (await firstComponent.decimals.callAsync()).toNumber();
+      kyberConversionRatePower = new BigNumber(10).pow(18 + sourceTokenDecimals - destinationTokenDecimals);
+
+      // NOTE: Kyber Minimum Conversion rates should be < 3.2 x 10**17
+      const minimumConversionRate = maxDestinationQuantity.div(sourceTokenQuantity)
+                                                          .mul(kyberConversionRatePower)
+                                                          .round();
+
+      kyberTrade = {
+        sourceToken: firstComponent.address,
+        destinationToken: receiveToken.address,
+        sourceTokenQuantity: sourceTokenQuantity,
+        minimumConversionRate: minimumConversionRate,
+        maxDestinationQuantity: maxDestinationQuantity,
+      } as KyberTrade;
+
+      // Create 0x order for the second component, using ether(4) sentToken as default
+      zeroExOrderTakerTokenAmount = zeroExOrderTakerTokenAmount || exchangeRedeemSentTokenAmounts[1];
+      zeroExOrder = await setUtils.generateZeroExSignedFillOrder(
+        NULL_ADDRESS,                                      // senderAddress
+        zeroExOrderMaker,                                  // makerAddress
+        NULL_ADDRESS,                                      // takerAddress
+        ZERO,                                              // makerFee
+        ZERO,                                              // takerFee
+        zeroExOrderMakerTokenAmount,                       // makerAssetAmount
+        zeroExOrderTakerTokenAmount,                       // takerAssetAmount
+        receiveToken.address,                              // makerAssetAddress
+        secondComponent.address,                           // takerAssetAddress
+        SetUtils.generateSalt(),                           // salt
+        SetTestUtils.ZERO_EX_EXCHANGE_ADDRESS,             // exchangeAddress
+        NULL_ADDRESS,                                      // feeRecipientAddress
+        SetTestUtils.generateTimestamp(10000),             // expirationTimeSeconds
+        zeroExOrderTakerTokenAmount,                       // amount of zeroExOrder to fill
+      );
+
+      subjectExchangeOrdersData =
+        subjectExchangeOrdersData || setUtils.generateSerializedOrders([zeroExOrder, kyberTrade]);
+
+      // Approve the receive token to the 0x maker
+      await erc20Wrapper.approveTransfersAsync(
+        [receiveToken],
+        SetTestUtils.ZERO_EX_ERC20_PROXY_ADDRESS,
+        zeroExOrderMaker
+      );
+
+      // Fund the 0x maker with the receive token
+      await erc20Wrapper.transferTokenAsync(
+        receiveToken,
+        zeroExOrderMaker,
+        zeroExOrderMakerTokenAmount,
+      );
+
+      await erc20Wrapper.approveTransfersAsync(
+        [firstComponent, secondComponent],
+        transferProxy.address,
+        contractDeployer
+      );
+
+      // Issue the Set and transfer to the caller
+      await coreWrapper.issueSetTokenAsync(
+        core,
+        setToken.address,
+        exchangeRedeemQuantity
+      );
+
+      await erc20Wrapper.transferTokenAsync(
+        setToken,
+        subjectCaller,
+        exchangeRedeemQuantity
+      );
+    });
+
+    afterEach(async () => {
+      exchangeRedeemQuantity = undefined;
+      exchangeRedeemReceiveTokens = undefined;
+      exchangeRedeemReceiveTokenAmounts = undefined;
+    });
+
+    async function subject(): Promise<string> {
+      return exchangeIssuanceModule.exchangeRedeem.sendTransactionAsync(
+        subjectExchangeIssuanceParams,
+        subjectExchangeOrdersData,
+        { from: subjectCaller, gas: DEFAULT_GAS },
+      );
+    }
+
+    it('redeems the correct quantity of the set for the sender', async () => {
+      const existingBalance = await setToken.balanceOf.callAsync(exchangeIssuanceCaller);
+
+      await subject();
+
+      await assertTokenBalanceAsync(setToken, existingBalance.sub(exchangeRedeemQuantity), exchangeIssuanceCaller);
+    });
+
+    it('increments the correct amount of Sent token', async () => {
+      const existingBalance = await receiveToken.balanceOf.callAsync(exchangeIssuanceCaller);
+
+      await subject();
+
+      const expectedNewBalance = existingBalance.add(totalReceiveToken);
+      const newBalance = await receiveToken.balanceOf.callAsync(exchangeIssuanceCaller);
+
+      await expect(newBalance).to.be.bignumber.equal(expectedNewBalance);
+    });
+
+    it('emits correct LogExchangeRedeem event', async () => {
+      const txHash = await subject();
+
+      const formattedLogs = await setTestUtils.getLogsFromTxHash(txHash);
+      const expectedLogs = LogExchangeRedeem(
+        setToken.address,
+        subjectCaller,
+        exchangeRedeemQuantity,
+        exchangeRedeemReceiveTokens,
+        exchangeRedeemReceiveTokenAmounts,
+        exchangeIssuanceModule.address
+      );
+
+      await SetTestUtils.assertLogEquivalence(formattedLogs, expectedLogs);
+    });
+
+    describe('when a sentToken is not a component of the Set', async () => {
+      before(async () => {
+        const firstComponent = erc20Wrapper.kyberReserveToken(SetTestUtils.KYBER_RESERVE_SOURCE_TOKEN_ADDRESS);
+        const notComponent = await erc20Wrapper.deployTokenAsync(contractDeployer);
+
+        exchangeRedeemSentTokens = [firstComponent.address, notComponent.address];
+      });
+
+      after(async () => {
+        exchangeRedeemSentTokens = undefined;
+      });
+
+      it('should revert', async () => {
+        await expectRevertError(subject());
+      });
+    });
+
+    describe('when quantity is zero', async () => {
+      before(async () => {
+        exchangeRedeemQuantity = new BigNumber(0);
+      });
+
+     after(async () => {
+        exchangeRedeemQuantity = undefined;
+      });
+
+      it('should revert', async () => {
+        await expectRevertError(subject());
+      });
+    });
+
+    describe('when sent quantities is zero', async () => {
+      before(async () => {
+        exchangeRedeemSentTokenAmounts = [ZERO, new BigNumber(0)];
+      });
+
+     after(async () => {
+        exchangeRedeemSentTokenAmounts = undefined;
+      });
+
+      it('should revert', async () => {
+        await expectRevertError(subject());
+      });
+    });
+
+    describe('when there are fewer receive tokens than expected', async () => {
+      before(async () => {
+        exchangeRedeemReceiveTokenAmounts = [totalReceiveToken.times(2)];
+      });
+
+     after(async () => {
+        exchangeRedeemReceiveTokenAmounts = undefined;
       });
 
       it('should revert', async () => {
