+++ test/contracts/supplementary/payableExchangeIssue.spec.ts
@@ -4,13 +4,13 @@ import * as ABIDecoder from 'abi-decoder';
 import * as chai from 'chai';
 import { BigNumber } from 'bignumber.js';
 import * as setProtocolUtils from 'set-protocol-utils';
-import { Address, Bytes, ExchangeInteractData, ZeroExSignedFillOrder } from 'set-protocol-utils';
+import { Address, Bytes, ExchangeIssuanceParams, ZeroExSignedFillOrder } from 'set-protocol-utils';
 
 import ChaiSetup from '@utils/chaiSetup';
 import { BigNumberSetup } from '@utils/bigNumberSetup';
 import {
   CoreContract,
-  ExchangeIssueModuleContract,
+  ExchangeIssuanceModuleContract,
   PayableExchangeIssueContract,
   RebalancingSetTokenContract,
   RebalancingSetTokenFactoryContract,
@@ -55,7 +55,7 @@ contract('PayableExchangeIssue', accounts => {
   ] = accounts;
 
   let core: CoreContract;
-  let exchangeIssueModule: ExchangeIssueModuleContract;
+  let exchangeIssuanceModule: ExchangeIssuanceModuleContract;
   let transferProxy: TransferProxyContract;
   let vault: VaultContract;
   let rebalancingSetTokenFactory: RebalancingSetTokenFactoryContract;
@@ -89,15 +89,15 @@ contract('PayableExchangeIssue', accounts => {
     rebalancingSetTokenFactory = await coreWrapper.deployRebalancingSetTokenFactoryAsync(core.address, whitelist);
     await coreWrapper.addFactoryAsync(core, rebalancingSetTokenFactory);
 
-    exchangeIssueModule = await coreWrapper.deployExchangeIssueModuleAsync(core, vault);
-    await coreWrapper.addModuleAsync(core, exchangeIssueModule.address);
+    exchangeIssuanceModule = await coreWrapper.deployExchangeIssuanceModuleAsync(core, vault);
+    await coreWrapper.addModuleAsync(core, exchangeIssuanceModule.address);
 
     weth = await erc20Wrapper.deployWrappedEtherAsync(ownerAccount);
 
     payableExchangeIssue = await payableExchangeIssueWrapper.deployPayableExchangeIssueAsync(
       core.address,
       transferProxy.address,
-      exchangeIssueModule.address,
+      exchangeIssuanceModule.address,
       weth.address,
     );
 
@@ -130,7 +130,7 @@ contract('PayableExchangeIssue', accounts => {
       return await payableExchangeIssueWrapper.deployPayableExchangeIssueAsync(
         core.address,
         transferProxy.address,
-        exchangeIssueModule.address,
+        exchangeIssuanceModule.address,
         weth.address,
         subjectCaller,
       );
@@ -152,19 +152,19 @@ contract('PayableExchangeIssue', accounts => {
       expect(coreAddress).to.equal(core.address);
     });
 
-    it('should contain the correct address of the exchangeIssueModule', async () => {
+    it('should contain the correct address of the exchangeIssuanceModule', async () => {
       const payableExchangeIssueContract = await subject();
 
-      const exchangeIssueModuleAddress = await payableExchangeIssueContract.exchangeIssueModule.callAsync();
+      const exchangeIssuanceModuleAddress = await payableExchangeIssueContract.exchangeIssuanceModule.callAsync();
 
-      expect(exchangeIssueModuleAddress).to.equal(exchangeIssueModule.address);
+      expect(exchangeIssuanceModuleAddress).to.equal(exchangeIssuanceModule.address);
     });
   });
 
   describe('#issueRebalancingSetWithEther', async () => {
     const subjectCaller: Address = tokenPurchaser;
     let subjectRebalancingSetAddress: Address;
-    let subjectExchangeInteractData: ExchangeInteractData;
+    let subjectExchangeIssuanceParams: ExchangeIssuanceParams;
     let subjectExchangeOrdersData: Bytes;
     let subjectEther: BigNumber;
 
@@ -229,7 +229,7 @@ contract('PayableExchangeIssue', accounts => {
         unit => unit.mul(exchangeIssueQuantity).div(baseSetNaturalUnit)
       );
 
-      subjectExchangeInteractData = {
+      subjectExchangeIssuanceParams = {
         setAddress: 			      exchangeIssueSetAddress,
         sentTokenExchangeIds:     exchangeIssueSentTokenExchangeIds,
         sentTokens:             exchangeIssueSentTokens,
@@ -277,7 +277,7 @@ contract('PayableExchangeIssue', accounts => {
     async function subject(): Promise<string> {
       return payableExchangeIssue.issueRebalancingSetWithEther.sendTransactionAsync(
         subjectRebalancingSetAddress,
-        subjectExchangeInteractData,
+        subjectExchangeIssuanceParams,
         subjectExchangeOrdersData,
         {
           from: subjectCaller,
